requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
jwt==1.3.1
passlib==1.7.4
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2

python-jose~=3.4.0
aiofiles~=24.1.0
pillow_heif~=0.22.0
openai~=1.78.0
opencv-python~=4.10.0.84
supervision~=0.25.1